delete process commands
